Fix compiler warnings (output unchanged)
- Fix nested comments (-Wcomment)
- Rename GREEK to GREEKFONT to avoid macro conflict
- Replace gets() with fgets() (-Wdeprecated-declarations)
- Fix format specifiers (-Wformat)
- Fix unused value expression
- Remove obsolete header files

Warnings reduced from 1046 to 775.

+++ src/morphlib/endio.c
@@ -164,7 +164,7 @@ get_endheader(FILE *f, int *maxp)
 #else
 	if( endlen % (long) unitsize ) {
 #endif
-printf("gstrsize %d endlen %d unitsize %d, mod %d, nendings %d filelen %ld\n", gstrsize, endlen, unitsize, ( endlen % unitsize ), nendings, filelen );
+printf("gstrsize %d endlen %d unitsize %d, mod %d, nendings %d filelen %ld\n", gstrsize, endlen, unitsize, ( endlen % unitsize ), nendings, (long)filelen );
 		fprintf(stderr,"Error in endio!\n");
 		return(0);
 	}
